# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Approval Flow API",
    "description": "Public REST API for Smart Approval Flow. All responses use a consistent envelope. See docs/api-design.md.",
    "version": "1.0.0-block1"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session JWT (users) or API key (integrations, Block 5)."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/approval-cases": {
      "get": {
        "summary": "List approval cases for the current tenant",
        "tags": [
          "approval-cases"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Create an approval case",
        "tags": [
          "approval-cases"
        ],
        "description": "Validated against the shared Zod schema. Supports the `Idempotency-Key` header to make retries safe.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/approval-cases/{id}": {
      "get": {
        "summary": "Get one approval case with items, decisions and audit trail",
        "tags": [
          "approval-cases"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/approval-cases/{id}/generate-public-link": {
      "post": {
        "summary": "Issue (or rotate) the secure loginless customer link",
        "tags": [
          "approval-cases"
        ],
        "description": "Creates a fresh access token, marks the case as SENT (pending customer), writes an audit event and returns the full customer URL. The raw token is returned once and never stored in plaintext.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/approval-cases/{id}/send": {
      "post": {
        "summary": "Send the approval request (issues link + e-mails the customer)",
        "tags": [
          "approval-cases"
        ],
        "description": "Rotates a fresh secure link, e-mails the customer using the initial-request template, moves the case to SENT and records the send in the audit trail and outbound messages. Emits approval_case.sent.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/approval-cases/{id}/remind": {
      "post": {
        "summary": "Send a reminder for a pending case",
        "tags": [
          "approval-cases"
        ],
        "description": "Only allowed while a customer response is still expected (SENT/VIEWED/CALLBACK). Rotates the link, e-mails the reminder template, increments the reminder counter and records it. Emits approval_case.reminder_sent.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/approval-cases/{id}/timeline": {
      "get": {
        "summary": "Merged, chronological history of a case (audit + messages)",
        "tags": [
          "approval-cases"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
}